Fixed end-game menu
Now shows whether the player won or lost, and the options for "play again" and "main menu" now function.  Only glitch is that when you press "play again" the wizard casts two spells because of "wrds.update()" being called.... not sure how to fix that but also not a huge issue.